debt - extract layout from workbench
